+++ src/main/java/com/drmangotea/tfmg/content/electricity/generators/creative_generator/CreativeGeneratorBlockEntity.java
@@ -56,7 +56,7 @@ public class CreativeGeneratorBlockEntity extends ElectricBlockEntity {
     }
 
     @Override
-    public int powerGeneration() {
+    public float powerGeneration() {
         return 999999999;
     }
