+++ src/main/java/com/drmangotea/tfmg/content/machinery/vat/freezer/FreezerBlockEntity.java
@@ -71,7 +71,7 @@ public class FreezerBlockEntity extends ElectricBlockEntity  {
 
 
     @Override
-    public void onNetworkChanged(int oldVoltage, int oldPower) {
+    public void onNetworkChanged(int oldVoltage, float oldPower) {
         super.onNetworkChanged(oldVoltage, oldPower);
         VatBlock.updateVatState(getBlockState(), level, getBlockPos().relative(Direction.DOWN));
     }
